gradio_frontend/requirements.txt
gradio>=4.36.1
requests>=2.32.3
python-dotenv>=1.0.1
huggingface-hub>=0.23.0